refactor accordions
